@@ -6,7 +6,7 @@ description: Demonstrates how to use the vdotok_connect plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -20,24 +20,24 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  scroll_to_index: ^2.0.0
+  scroll_to_index: ^3.0.1
   cupertino_icons: ^1.0.2
-  page_transition: ^1.1.7+3
-  shared_preferences:
-  http: ^0.13.0
-  json_annotation: ^3.1.1
+  page_transition: ^2.0.5
+  shared_preferences: ^2.0.5
+  http: ^0.13.4
+  json_annotation: ^4.5.0
   provider: ^5.0.0
   device_info: ^2.0.0
   flutter_icons: ^1.1.0
-  image_picker: ^0.7.5+3
+  image_picker: ^0.8.5+3
   path: ^1.8.0
-  path_provider:
+  path_provider: ^2.0.1
   audioplayers:
   data_connection_checker: ^0.3.4
   email_validator: ^2.0.1
   flutter_svg: ^0.22.0
   intl: ^0.17.0
-  file_picker: ^3.0.2+2
+  file_picker: ^4.6.1
   #video_player: ^1.0.1
   # chewie: ^0.12.2
   open_file: ^3.2.1
