@@ -24,17 +24,17 @@ class _SignUpScreenState extends State<SignUpScreen> {
   bool emailvalidate = false;
   bool passwordvalidate = false;
   bool namevalidate = false;
-  Size size;
+  Size? size;
 
   handlePress() async {
-    if (_registerformkey.currentState.validate()) {
+    if (_registerformkey.currentState!.validate()) {
       AuthProvider auth = Provider.of<AuthProvider>(context, listen: false);
       bool res = await auth.register(
         _nameController.text,
         _passwordController.text,
         _emailController.text,
       );
-      if (auth.getUser.auth_token == null) {
+      if (auth.getUser!.auth_token == null) {
         setState(() {
           _autoValidate = true;
         });
@@ -60,7 +60,7 @@ class _SignUpScreenState extends State<SignUpScreen> {
     ));
 
     size = MediaQuery.of(context).size;
-    print("The size is : ${size.height * 1.08}");
+    print("The size is : ${size!.height * 1.08}");
     return GestureDetector(
         onTap: () {
           FocusScopeNode currentFous = FocusScope.of(context);
