@@ -4,7 +4,7 @@ import '../constants/constant.dart';
 class PasswordFieldFile extends StatefulWidget {
   String name;
   var myController = TextEditingController();
-  PasswordFieldFile({Key key, this.name, this.myController}) : super(key: key);
+  PasswordFieldFile({Key? key, required this.name, required this.myController}) : super(key: key);
   @override
   _WidgetHeaderState createState() => _WidgetHeaderState();
 }
@@ -28,7 +28,7 @@ class _WidgetHeaderState extends State<PasswordFieldFile> {
         obscureText: true,
         controller: myController,
         validator: (value) {
-          if (value.isEmpty) return "Field cannot be empty";
+          if (value!.isEmpty) return "Field cannot be empty";
           if (value.length < 6)
             return "Entry should be at least 6 characters long";
           if (value.length > 14) return "Entry should not exceed 14 characters";
