@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import '../constants/constant.dart';
 
 class SignInButtonFile extends StatefulWidget {
-  String name;
+  String ?name;
   final handlePress;
   bool _autoValidate = true;
   var myController = TextEditingController();
-  SignInButtonFile({Key key, this.name, this.handlePress}) : super(key: key);
+  SignInButtonFile({Key? key, this.name, this.handlePress}) : super(key: key);
   @override
   _WidgetHeaderState createState() => _WidgetHeaderState();
 }
