@@ -3,13 +3,13 @@ part 'contact.g.dart';
 
 @JsonSerializable()
 class Contact {
-  int user_id;
-  dynamic email;
-  String ref_id;
+  int ?user_id;
+  dynamic? email;
+  String? ref_id;
   String full_name;
-  bool isSelected = false;
+  bool ?isSelected = false;
 
-  Contact({this.user_id, this.ref_id, this.full_name,this.email, this.isSelected = false});
+  Contact({this.user_id, this.ref_id, required this.full_name,this.email, this.isSelected = false});
   factory Contact.fromJson(Map<String, dynamic> json) =>
       _$ContactFromJson(json);
   Map<String, dynamic> toJson() => _$ContactToJson(this);
