@@ -6,23 +6,23 @@ class User {
   //final String active;
   final String auth_token;
   final String authorization_token;
-  final String email;
+  final String ?email;
   final String full_name;
-  final String message;
+  final String ?message;
   //final String passwordCount;
-  final int process_time;
+  final int? process_time;
   final String ref_id;
-  final int status;
-  final int user_id;
+  final int ?status;
+  final int ?user_id;
   User(
       {
-      this.auth_token,
-      this.authorization_token,
+      required this.auth_token,
+      required this.authorization_token,
       this.email,
-      this.full_name,
+      required this.full_name,
       this.message,
       this.process_time,
-      this.ref_id,
+      required this.ref_id,
       this.status,
       this.user_id,
      });
