@@ -7,9 +7,9 @@ import '../jsManager/jsManagerStub.dart'
     if (dart.library.html) '../jsManager/jsManagerForWeb.dart';
 
 abstract class JsManager {
-  static JsManager _instance;
+  static JsManager? _instance;
 
-  static JsManager get instance {
+  static JsManager? get instance {
     _instance ??= getJsManager();
     return _instance;
   }
