@@ -4,10 +4,11 @@
 
 // ignore_for_file: directives_ordering
 // ignore_for_file: lines_longer_than_80_chars
+// ignore_for_file: depend_on_referenced_packages
 
 import 'package:audioplayers/web/audioplayers_web.dart';
 import 'package:connectivity_plus_web/connectivity_plus_web.dart';
-import 'package:file_picker/src/file_picker_web.dart';
+import 'package:file_picker/_internal/file_picker_web.dart';
 import 'package:fluttertoast/fluttertoast_web.dart';
 import 'package:image_picker_for_web/image_picker_for_web.dart';
 import 'package:shared_preferences_web/shared_preferences_web.dart';
