@@ -26,10 +26,10 @@ class _RegisterScreenState extends State<RegisterScreen> {
   }
 
   handlePress() async {
-    if (_registerformkey.currentState.validate()) {
+    if (_registerformkey.currentState!.validate()) {
       AuthProvider auth = Provider.of<AuthProvider>(context, listen: false);
       bool res = await auth.register(_emailController.text, _passwordController.text,"");
-      if (auth.getUser.auth_token == null || auth.getUser.auth_token.isEmpty) {
+      if (auth.getUser!.auth_token == null || auth.getUser!.auth_token.isEmpty) {
         setState(() {
           _autoValidate = true;
         });
