@@ -5,7 +5,7 @@ class ReusableButton extends StatefulWidget {
   final String text;
   final handlePress;
 
-  ReusableButton({this.text, this.handlePress});
+  ReusableButton({required this.text, this.handlePress});
 
   @override
   _ReusableButtonState createState() => _ReusableButtonState();
