@@ -12,7 +12,7 @@ Contact _$ContactFromJson(Map<String, dynamic> json) {
     ref_id: json['ref_id'] as String,
     full_name: json['full_name'] as String,
     email: json['email'],
-    isSelected: json['isSelected'] as bool,
+    isSelected: json['isSelected'],
   );
 }
 
